refactor: move tables to export file
